Create working rma echo program
The code isn't super clean, but it works. I don't free any resources,
but I may add that in a later update. Another issue is that the server's
counter doesn't get incremented after its write which prevents the
server code from finishing. It still sends the message however, so the
program works.